Gen AI Apps/Writing Assistant/requirements.txt
streamlit
langchain
python-dotenv
langchain_groq